+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Hitting a mob more than 40 blocks away with an arrow now plays the ding sound
 * Timber Trees now breaks some leaves before falling (35% for Quark's Trumpet trees, 10% the others)
+* Campfires are now extinguished when it starts to rain
 * Changed Efficiency and Unbreaking formulas
   * Efficiency bonus: ~~+12/30/60/102/156/222%~~ -> +25% per level 
   * Unbreaking: tool lasts ~~40/115/190/265/340/415%~~ -> 50% per level more